src/main/java/faang/school/projectservice/filter/ProjectStatusFilter.java
package faang.school.projectservice.filter;

import faang.school.projectservice.dto.filter.ProjectFilterDto;
import faang.school.projectservice.model.Project;
import org.springframework.stereotype.Component;

import java.util.regex.Pattern;
import java.util.stream.Stream;

@Component
public class ProjectStatusFilter implements ProjectFilter {

    @Override
    public boolean isApplicable(ProjectFilterDto filters) {
        return filters.statusPattern() != null;
    }

    @Override
    public Stream<Project> apply(Stream<Project> projects, ProjectFilterDto filters) {
        var pattern = Pattern.compile(Pattern.quote(filters.statusPattern()), Pattern.CASE_INSENSITIVE);
        return projects.filter(project -> pattern.matcher(project.getStatus().name()).find());
    }
}
